Extract common utility functions.
